python-flask-google-login/requirements.txt
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
google-auth==2.29.0
google-auth-oauthlib==1.2.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
